Release: Merge develop into main - 2025-04-21 (#3)
* fix: update author and readTime in blog post metadata

Corrected the author name and read time for the "Simplifying Development with Dev Containers" blog post to ensure accurate information is presented.

* fix: update readTime in blog post metadata

Adjusted the readTime for the blog post on Dev Containers from '8 min read' to '10 min read' for accuracy.

* chore: install packages

* chore: delete yarn lock

* feat: enhance project configuration

- Updated .gitignore to include .sst and .open-next directories
- Added sst-env.d.ts for SST resource definitions
- Created sst.config.ts for SST application configuration
- Modified tsconfig.json to exclude sst.config.ts from compilation

* feat: enhance SEO and metadata for portfolio

Updated metadata for improved SEO and added robots and sitemap functionality.
- Enhanced title and description in metadata
- Added robots.txt for search engine indexing
- Implemented sitemap generation for better site navigation
- Introduced JSON-LD structured data for improved search visibility

* feat: update GitHub link in Projects component

Changed the GitHub link in the Projects component to point to the correct user profile.